refactor: extracted DictValue into dict_value.py
